fix joomla

+++ scripts/cleanup.sh
@@ -1,9 +1,16 @@
-#!/bin/bash -ex
+#!/bin/sh -e
 
-# Make sure udev does not block our network - http://6.ptmc.org/?p=164
+cat <<EOF > /etc/resolv.conf
+#nameserver 2001:4860:4860::8888
+#nameserver 2001:4860:4860::8844
+nameserver 8.8.8.8
+nameserver 8.8.4.4
+options timeout:2 attempts:1 rotate
+EOF
+
 echo "==> Cleaning up udev rules"
 rm -rf /dev/.udev/
-rm /lib/udev/rules.d/75-persistent-net-generator.rules
+test -f /lib/udev/rules.d/75-persistent-net-generator.rules && rm /lib/udev/rules.d/75-persistent-net-generator.rules
 
 echo "==> Cleaning up leftover dhcp leases"
 # Ubuntu 10.04
@@ -13,27 +20,22 @@ fi
 # Ubuntu 12.04 & 14.04
 if [ -d "/var/lib/dhcp" ]; then
     rm /var/lib/dhcp/*
-fi 
+fi
 
-echo "==> Cleaning up tmp"
 rm -rf /tmp/*
 
 uname -a
 export APT_LISTCHANGES_FRONTEND=none
-DEBIAN_FRONTEND=noninteractive apt-get -o Dpkg::Options::="--force-confnew" --force-yes -fuy purge $(dpkg --list | grep '^rc' |awk '{print $2}')
-DEBIAN_FRONTEND=noninteractive apt-get -o Dpkg::Options::="--force-confnew" --force-yes -fuy purge $(dpkg --list | awk -v image="$(uname -r)" '/linux-image-[0-9]/{if($0 !~ image) print $2 }')
-DEBIAN_FRONTEND=noninteractive apt-get -o Dpkg::Options::="--force-confnew" --force-yes -fuy autoremove --purge
-DEBIAN_FRONTEND=noninteractive apt-get -o Dpkg::Options::="--force-confnew" --force-yes -y clean
-DEBIAN_FRONTEND=noninteractive apt-get -o Dpkg::Options::="--force-confnew" --force-yes -y autoclean
-
+export DEBIAN_FRONTEND=noninteractive
+apt-get -o Dpkg::Options::="--force-confnew" --force-yes -fuy purge $(dpkg --list | grep '^rc' |awk '{print $2}')
+apt-get -o Dpkg::Options::="--force-confnew" --force-yes -fuy purge $(dpkg --list | awk -v image="$(uname -r)" '/linux-image-[0-9]/{if($0 !~ image) print $2 }')
+apt-get -o Dpkg::Options::="--force-confnew" --force-yes -fuy autoremove --purge
+apt-get -o Dpkg::Options::="--force-confnew" --force-yes -y clean
+apt-get -o Dpkg::Options::="--force-confnew" --force-yes -y autoclean
 
 update-initramfs -u
 update-grub
 
-echo "==> Installed packages"
-dpkg --get-selections | grep -v deinstall
-
-# Remove Bash history
 unset HISTFILE
 rm -f /root/.bash_history
 rm -f /home/vagrant/.bash_history
@@ -41,62 +43,14 @@ rm -f /home/vagrant/.bash_history
 # Clean up log files
 find /var/log -type f | while read f; do echo -ne '' > $f; done;
 
-# Make sure we wait until all the data is written to disk, otherwise
-# Packer might quite too early before the large files are deleted
-sync
-
-
-echo "==> Installed packages before cleanup"
-dpkg --get-selections | grep -v deinstall
-
-# Remove some packages to get a minimal install
-echo "==> Removing all linux kernels except the currrent one"
-dpkg --list | awk '{ print $2 }' | grep 'linux-image-3.*-generic' | grep -v $(uname -r) | xargs apt-get -y purge
-echo "==> Removing linux headers"
-dpkg --list | awk '{ print $2 }' | grep linux-headers | xargs apt-get -y purge
-rm -rf /usr/src/linux-headers*
-echo "==> Removing linux source"
-dpkg --list | awk '{ print $2 }' | grep linux-source | xargs apt-get -y purge
-echo "==> Removing development packages"
-dpkg --list | awk '{ print $2 }' | grep -- '-dev$' | xargs apt-get -y purge
-echo "==> Removing documentation"
-dpkg --list | awk '{ print $2 }' | grep -- '-doc$' | xargs apt-get -y purge
-echo "==> Removing development tools"
-#dpkg --list | grep -i compiler | awk '{ print $2 }' | xargs apt-get -y purge
-#apt-get -y purge cpp gcc g++ 
-apt-get -y purge build-essential
-echo "==> Removing default system Ruby"
-#apt-get -y purge ruby ri doc
-echo "==> Removing default system Python"
-#apt-get -y purge python-dbus libnl1 python-smartpm python-twisted-core libiw30 python-twisted-bin libdbus-glib-1-2 python-pexpect python-pycurl python-serial python-gobject python-pam python-openssl
-echo "==> Removing X11 libraries"
-#apt-get -y purge libx11-data xauth libxmuu1 libxcb1 libx11-6 libxext6
-echo "==> Removing obsolete networking components"
-apt-get -y purge ppp pppconfig pppoeconf
-echo "==> Removing other oddities"
-apt-get -y purge popularity-contest installation-report landscape-common wireless-tools wpasupplicant
-
-# Clean up the apt cache
 apt-get -y autoremove --purge
 apt-get -y autoclean
 apt-get -y clean
 
-# Clean up orphaned packages with deborphan
-apt-get -y install deborphan
-while [ -n "$(deborphan --guess-all --libdevel)" ]; do
-    deborphan --guess-all --libdevel | xargs apt-get -y purge
+for p in ufw ntfs-3g netcat-openbsd 'language-pack-gnome-*' ureadahead rsyslog tcpd accountsservice install-info  krb5-locales laptop-detect lshw mlocate ntpdate command-not-found-data powermgmt-base build-essential ppp pppconfig pppoeconf popularity-contest installation-report landscape-common wireless-tools wpasupplicant; do
+    apt-get -y purge $p || true;
 done
-apt-get -y purge deborphan dialog
 
-echo "==> Removing man pages"
-rm -rf /usr/share/man/*
-echo "==> Removing APT files"
-find /var/lib/apt -type f | xargs rm -f
-echo "==> Removing anything in /usr/src"
-rm -rf /usr/src/*
-echo "==> Removing any docs"
-rm -rf /usr/share/doc/*
-echo "==> Removing caches"
-find /var/cache -type f | xargs rm -f
+find /var/lib/apt/lists/ -type f -delete
 
-fstrim -v /
+fstrim -v / || echo dummy
